Add top-level configure/Makefile
This takes care to build the sub-projects in the correct order and with
the correct environment variables set such that downstream projects will
include/link within these project rather than against system-installed
versions.

`configure` is a Python `Makefile` generator rather than autotools.  This
has made it a little easier to implement.  The design is influenced by
[ninja]/[meson]: the generated `Makefile` is very simple and contains no
pattern rules or conditionals.  Any complexity is restricted to the
`configure` script.  This is faster and much easier to debug than using
advanced GNU Make features.

[ninja]: https://ninja-build.org/
[meson]: http://mesonbuild.com/