gst-uninstalled: Add option `--generate=makefile`
This is used to generate a Makefile fragment that will set all the
appropriate environment variables correctly.  It can be used by adding a
rule like:

    gst-uninstalled.mk : gstreamer/scripts/gst-uninstalled
        gstreamer/scripts/gst-uninstalled --generate=makefile >gst-uninstalled.mk

    include gst-uninstalled.mk

to your `Makefile`.  That way any rules that get run in the Makefile will
have the correct environment variables set, without requiring the caller
to have run `gst-uninstalled` themselves.